requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.2.2
imbalanced-learn>=0.10
joblib>=1.2.0
pyyaml>=6.0
shap>=0.41.0
streamlit>=1.20.0
matplotlib>=3.6
pytest>=7.0.0
pytest-cov>=4.0.0
flake8>=5.0.0
black>=22.0.0